services/notification/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic[email]
motor==3.4.0
pymongo==4.5.0